requirements.txt
streamlit==1.31.0
pandas>=2.0.3,<3.0.0
pydantic>=2.0.3,<3.0.0
# Requires system-level libGL dependency
opencv-python>=4.8.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0,<7.0
pydantic-settings>=2.10.1,<3.0.0
openai>=1.98.0,<2.0.0
aiohttp>=3.9.5,<4.0.0
streamlit-drawable-canvas-fix==0.9.7
gitpython<3.1.19
altair<6
pyarrow<21.0.0
